linux: move package to root
It doesn't really make sense for an "implementer" to exist inside the
indexer/ dir so in the spirit of cleaning up indexer/ move the linux
specific package.
