+++ dpkg/ecosystem.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/quay/claircore/debian"
 	"github.com/quay/claircore/indexer"
-	"github.com/quay/claircore/indexer/linux"
+	"github.com/quay/claircore/linux"
 	"github.com/quay/claircore/ubuntu"
 )
