+++ linux/distsearcher.go
