+++ linux/packagesearcher.go
