+++ rpm/ecosystem.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/quay/claircore/aws"
 	"github.com/quay/claircore/indexer"
-	"github.com/quay/claircore/indexer/linux"
+	"github.com/quay/claircore/linux"
 	"github.com/quay/claircore/oracle"
 	"github.com/quay/claircore/photon"
 	"github.com/quay/claircore/suse"
